requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
apscheduler
anthropic
python-telegram-bot
httpx
aiohttp
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
loguru
